print "?" in place of non-printing bytes of obj/key/val values
* tools/corosync-objctl.c: Include <ctype.h>
(print_name): New function.
(tail_key_changed): Use it to avoid printing garbage to screen.
* include/corosync/confdb.h (confdb_key_change_notify_fn_t): Convert
type of "int" length params to "size_t".

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2000 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ include/corosync/confdb.h
@@ -63,11 +63,11 @@ typedef void (*confdb_key_change_notify_fn_t) (
 	hdb_handle_t parent_object_handle,
 	hdb_handle_t object_handle,
 	const void *object_name,
-	int  object_name_len,
+	size_t  object_name_len,
 	const void *key_name,
-	int key_name_len,
+	size_t key_name_len,
 	const void *key_value,
-	int key_value_len);
+	size_t key_value_len);
 
 typedef void (*confdb_object_create_notify_fn_t) (
 	confdb_handle_t handle,
